Use IntelliJ to run home.html (built-in Server)
Only image fetching is left.
Also, when you open home.html, the contents won't be displayed by normally opening the browser. The reason being we are not yet allowed access to collegespace or nsitpedia feeds. That's y there are sattic XML files saved since we're not on server right now.
These static files also need web server to be accessed.
So, for now these can't be viewed but they'll be there when we put it on server.
So, notepad or normal browser opening won't display the contents since these don't have server right now.
Although you can open it in intelliJ ide, since it's recent editions come with in-built Web Server.